+++ src/core.rs
@@ -1,16 +1,19 @@
 use std::collections::HashMap;
 
+use lazy_static::lazy_static;
+use regex::{Captures, Regex};
+
 use crate::printer::pr_str;
 use crate::types::{Args, Function, Ret, Type};
 
 pub struct Namespace {
-    data: HashMap<String, Function>
+    data: HashMap<String, Function>,
 }
 
 impl Namespace {
     pub fn new() -> Namespace {
         Namespace {
-            data: HashMap::new()
+            data: HashMap::new(),
         }
     }
 
@@ -20,7 +23,10 @@ impl Namespace {
         ns.data.insert(String::from("-"), sub);
         ns.data.insert(String::from("*"), mul);
         ns.data.insert(String::from("/"), div);
+        ns.data.insert(String::from("pr-str"), pr_str_fun);
+        ns.data.insert(String::from("str"), str_fun);
         ns.data.insert(String::from("prn"), prn);
+        ns.data.insert(String::from("println"), println);
         ns.data.insert(String::from("list"), list);
         ns.data.insert(String::from("list?"), is_list);
         ns.data.insert(String::from("empty?"), is_empty);
@@ -49,7 +55,9 @@ fn sum(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a + *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 + *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a + *b)),
-        _ => Err(String::from("Type error: '+' is only supported for Int and Float")),
+        _ => Err(String::from(
+            "Type error: '+' is only supported for Int and Float",
+        )),
     }
 }
 
@@ -59,7 +67,9 @@ fn sub(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a - *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 - *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a - *b)),
-        _ => Err(String::from("Type error: '-' is only supported for Int and Float")),
+        _ => Err(String::from(
+            "Type error: '-' is only supported for Int and Float",
+        )),
     }
 }
 
@@ -69,7 +79,9 @@ fn mul(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a * *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 * *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a * *b)),
-        _ => Err(String::from("Type error: '*' is only supported for Int and Float")),
+        _ => Err(String::from(
+            "Type error: '*' is only supported for Int and Float",
+        )),
     }
 }
 
@@ -79,18 +91,86 @@ fn div(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a / *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 / *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a / *b)),
-        _ => Err(String::from("Type error: '/' is only supported for Int and Float")),
+        _ => Err(String::from(
+            "Type error: '/' is only supported for Int and Float",
+        )),
     }
 }
 
+fn pr_str_fun(args: Args) -> Ret {
+    let s = args
+        .iter()
+        .map(|arg| pr_str(arg.clone(), true))
+        .collect::<Vec<String>>()
+        .join(" ");
+
+    Ok(Type::String(s))
+}
+
+fn str_fun(args: Args) -> Ret {
+    let s = args
+        .iter()
+        .map(|arg| {
+            let mut s = pr_str(arg.clone(), false);
+            lazy_static! {
+                static ref RE: Regex = Regex::new(r#"(["][^"])|([^"]["])"#).unwrap();
+            }
+            s = RE
+                .replace_all(&s, |cap: &Captures| match &cap[0] {
+                    _ if cap[0].starts_with('"') => cap[0].replacen("\"", "", 1),
+                    _ if cap[0].ends_with('"') => cap[0].replacen("\"", "", 1),
+                    _ => panic!("Impossible capture {}", &cap[0]),
+                })
+                .to_string();
+
+            if s.len() > 0 && s.starts_with('"') {
+                s[1..s.len() - 1].to_string()
+            } else {
+                s
+            }
+        })
+        .collect::<Vec<String>>()
+        .join("");
+
+    Ok(Type::String(s))
+}
+
 fn prn(args: Args) -> Ret {
-    match args.get(0) {
-        Some(value) => {
-            println!("{}", pr_str(value.clone(), true));
-            Ok(Type::Nil)
-        },
-        None => Err(format!("Must pass an argument to 'prn'")),
-    }
+    let s = args
+        .iter()
+        .map(|arg| pr_str(arg.clone(), true))
+        .collect::<Vec<String>>()
+        .join(" ");
+
+    println!("{}", s);
+    Ok(Type::Nil)
+}
+
+fn println(args: Args) -> Ret {
+    let s = args
+        .iter()
+        .map(|arg| {
+            let mut s = pr_str(arg.clone(), false);
+            let re = Regex::new(r#"(["][^"])|([^"]["])"#).unwrap();
+            s = re
+                .replace_all(&s, |cap: &Captures| match &cap[0] {
+                    _ if cap[0].starts_with('"') => cap[0].replacen("\"", "", 1),
+                    _ if cap[0].ends_with('"') => cap[0].replacen("\"", "", 1),
+                    _ => panic!("Impossible capture {}", &cap[0]),
+                })
+                .to_string();
+
+            if s.len() > 0 && s.starts_with('"') {
+                s[1..s.len() - 1].to_string()
+            } else {
+                s
+            }
+        })
+        .collect::<Vec<String>>()
+        .join(" ");
+
+    println!("{}", s);
+    Ok(Type::Nil)
 }
 
 fn list(args: Args) -> Ret {