+++ src/printer.rs
@@ -24,9 +24,27 @@ fn pr_type(t: &Type, print_readably: bool) -> String {
             } else {
                 s.push_str(&format!("{}.0", float))
             }
-        },
+        }
         Type::Symbol(symbol) => s.push_str(&format!("{}", symbol)),
         Type::Keyword(keyword) => s.push_str(&format!(":{}", keyword)),
+        Type::String(string) => {
+            let repr = if print_readably {
+                string
+                    .chars()
+                    .map(|c| match c {
+                        '"' => "\\\"".to_string(),
+                        '\\' => "\\\\".to_string(),
+                        '\n' => "\\n".to_string(),
+                        _ => c.to_string(),
+                    })
+                    .collect::<Vec<String>>()
+                    .join("")
+            } else {
+                string.clone()
+            };
+
+            s.push_str(&format!("\"{}\"", repr));
+        }
         Type::List(list) => {
             s.push('(');
             s.push_str(pr_seq(&list, print_readably).as_str());
@@ -44,8 +62,7 @@ fn pr_type(t: &Type, print_readably: bool) -> String {
 }
 
 fn pr_seq(seq: &Vec<Box<Type>>, print_readably: bool) -> String {
-    seq
-        .iter()
+    seq.iter()
         .map(|element| pr_type(&*element, print_readably))
         .collect::<Vec<String>>()
         .join(" ")